requirements.txt
beautifulsoup4>=4.14.3
chromadb>=1.5.2
fastapi>=0.135.1
notion-client>=3.0.0
openai>=2.26.0
pandas>=3.0.1
python-dotenv>=1.0.0
requests>=2.32.5
uvicorn>=0.41.0
playwright>=1.58.0